Extract and share Options
